KVM: svm: Do not install #UD intercept when SEV is active
On #UD, x86_emulate_instruction() fetches the data from guest memory and
decodes the instruction bytes to assist further. But when SEV is active,
the guest memory will be encrypted using the guest key hence we will not
able to fetch the instruction bytes from the guest memory to assist UD
handling.

Signed-off-by: Brijesh Singh <brijesh.singh@amd.com>