KVM: x86: Restart the guest when insn_len is zero and SEV is active
On AMD platform, under certain conditions insn_len may be zero on #NPF.
This can happen if guest gets a page-fault on data access, but HW table
walker is not able to read the instruction page (e.g instuction page
is not present in memory).

Typically, when insn_len is zero, x86_emulate_instruction() walks the
guest page table and fetches the instruction bytes from guest memory.
When SEV is active, the guest memory is encrypted with guest key hence
we will not able to fetch the instruction bytes. In those cases we
simply restart the guest.

I have encountered this issue when running kernbench inside the guest.

Signed-off-by: Brijesh Singh <brijesh.singh@amd.com>